Improved Alarm Trigger Function
Made several improvements to the alarm's trigger function. Notably, players can no longer set off their own alarms, trespassers can now be made wanted even if they are already wanted for setting off an alarm, the wanted message now specifies which door the tresspasser was trying to break into (if it has a name), and the function now has support for the SpectralRP base's wanted function if it is present.